cirrus 1.3 の変更を適用
mmload 命令のループ時間の型変更（ミリ秒（整数型）→秒（実数型））

dialog 命令一部実装
タイプ 0 ~ 3 （標準と警告の区別は無し）
タイプ 64 に入力ダイアログ実装（dialog "メッセージ",64,"デフォルトメッセージ" ;OKでrefstrに入力文字列が、キャンセルで空文字が返る）

devinfo res, "urlquery" ; res にURLクエリ取得
devprm "urlquery", "value" ; URLクエリを "value" に設定

コード整理

+++ dish.js/emscripten/1.34.1/system/include/SDL/SDL_mixer.h
@@ -535,7 +535,7 @@ extern DECLSPEC int SDLCALL Mix_GroupNewer(int tag);
 */
 #define Mix_PlayChannel(channel,chunk,loops,start,end) Mix_PlayChannelTimed(channel,chunk,loops,start,end)
 /* The same as above, but the sound is played at most 'ticks' milliseconds */
-extern DECLSPEC int SDLCALL Mix_PlayChannelTimed(int channel, Mix_Chunk *chunk, int loops, int start, int end);
+extern DECLSPEC int SDLCALL Mix_PlayChannelTimed(int channel, Mix_Chunk *chunk, int loops, double start, double end);
 extern DECLSPEC int SDLCALL Mix_PlayMusic(Mix_Music *music, int loops);
 
 /* Fade in music or a channel over "ms" milliseconds, same semantics as the "Play" functions */
